+++ processor/src/test/resources/bad_input/com/example/component/BadTypeComponent.java
@@ -2,7 +2,7 @@ package com.example.component;
 
 import arez.annotations.ArezComponent;
 
-@ArezComponent( type = "" )
+@ArezComponent( name = "" )
 public abstract class BadTypeComponent
 {
 }