+++ processor/src/test/resources/bad_input/com/example/component/BadTypeComponent2.java
@@ -2,7 +2,7 @@ package com.example.component;
 
 import arez.annotations.ArezComponent;
 
-@ArezComponent( type = "long" )
+@ArezComponent( name = "long" )
 public abstract class BadTypeComponent2
 {
 }